+++ inst/unitTests/test_voomWithDreamWeights.R
@@ -17,7 +17,12 @@ test_voomWithDreamWeights = function(){
 	vobj1 = voom( dge[1:100,], design)
 	vobj1$targets$sample.weights = 1
 
-	vobj2 = voomWithDreamWeights( dge[1:100,], form, data=metadata)
+	# original 
+	# vobj2 = voomWithDreamWeights( dge[1:100,], form, data=metadata)
+
+	# disable scaled pseudocounts
+	dge$counts = dge$counts + 0.5
+	vobj2 = voomWithDreamWeights( dge[1:100,], form, data=metadata, prior.count=0)
 
 	checkEquals(vobj1, vobj2)
 }
@@ -102,57 +107,64 @@ test_reweigthing_voom = function(){
 
 	library(edgeR)
 	library(limma)
+	library(variancePartition)
 	data(varPartDEdata)
 
 	# normalize RNA-seq counts
 	dge <- DGEList(counts = countMatrix)
 	dge <- calcNormFactors(dge)
 
+	# disable scaled pseudocounts
+	dge2 = dge
+	dge2$counts = dge2$counts + 0.5
+
 	form <- ~ Disease 
 	dsgn = model.matrix(form, metadata)
 
 	w = rep(2, nrow(metadata))
+	w.scale = w / mean(w)
 
 	# Recover voom results
 	#----------------------
 
 	# no weights
 	vobj1 = voom(dge, dsgn)
-	vobj2 <- voomWithDreamWeights(dge, form, metadata)
+	vobj2 <- voomWithDreamWeights(dge2, form, metadata, prior.count = 0)
 	checkEqualsNumeric(vobj1$weights, vobj2$weights)
 
 	# constant weights 
 	#-----------------
-	vobj1 = voom(dge, dsgn, weights=w)
+	vobj1 = voom(dge, dsgn, weights=w.scale)
 
 	# disable rescaling by input weights
-	vobj2 <- voomWithDreamWeights(dge, form, metadata, weights=w, rescaleWeightsAfter=FALSE)
+	vobj2 <- voomWithDreamWeights(dge2, form, metadata, weights=w.scale, rescaleWeightsAfter=FALSE, prior.count = 0)
 	checkEqualsNumeric(vobj1$weights, vobj2$weights)
 
 	# manual rescaling of weights after voom
-	vobj1 = voom(dge, dsgn, weights=w)
-	vobj1$weights <- t(w * t(vobj1$weights))
+	vobj1 = voom(dge, dsgn, weights=w.scale)
+	vobj1$weights <- t(w.scale * t(vobj1$weights))
 
 	# enble rescaling by input weights
-	vobj2 <- voomWithDreamWeights(dge, form, metadata, weights=w)
+	vobj2 <- voomWithDreamWeights(dge2, form, metadata, weights=w, prior.count = 0)
 	checkEqualsNumeric(vobj1$weights, vobj2$weights)
 
 	# varying weights 
 	#----------------
 	w = 1:nrow(metadata)
+	w.scale = w / mean(w)
 
-	vobj1 = voom(dge, dsgn, weights=w)
+	vobj1 = voom(dge, dsgn, weights=w.scale)
 
 	# disable rescaling by input weights
-	vobj2 <- voomWithDreamWeights(dge, form, metadata, weights=w, rescaleWeightsAfter=FALSE)
+	vobj2 <- voomWithDreamWeights(dge2, form, metadata, weights=w, rescaleWeightsAfter=FALSE, prior.count = 0)
 	checkEqualsNumeric(vobj1$weights, vobj2$weights)
 
 	# manual rescaling of weights after voom
-	vobj1 = voom(dge, dsgn, weights=w)
-	vobj1$weights <- t(w * t(vobj1$weights))
+	vobj1 = voom(dge, dsgn, weights=w.scale)
+	vobj1$weights <- t(w.scale * t(vobj1$weights))
 
 	# enble rescaling by input weights
-	vobj2 <- voomWithDreamWeights(dge, form, metadata, weights=w)
+	vobj2 <- voomWithDreamWeights(dge2, form, metadata, weights=w, prior.count = 0)
 	checkEqualsNumeric(vobj1$weights, vobj2$weights)
 
 	# weights as matrix
@@ -163,7 +175,7 @@ test_reweigthing_voom = function(){
 					ncol = length(w), 
 					byrow = TRUE)
 	# enble rescaling by input weights
-	vobj2 <- voomWithDreamWeights(dge, form, metadata, weights=weightsMatrix)
+	vobj2 <- voomWithDreamWeights(dge2, form, metadata, weights=weightsMatrix, prior.count = 0)
 	checkEqualsNumeric(vobj1$weights, vobj2$weights)
 }	
 
@@ -188,6 +200,10 @@ test_voomLmFit = function(){
 	b = getNormLibSizes(dge)
 	checkEqualsNumeric(a,b)
 
+	# disable scaled pseudocounts
+	dge2 = dge
+	dge2$counts = dge2$counts + 0.5
+
 	form <- ~ Disease 
 	dsgn = model.matrix(form, metadata)
 
@@ -197,7 +213,7 @@ test_voomLmFit = function(){
 	fit1 = voomLmFit2( dge, dsgn, prior.weights=w)
 	fit1$EList$genes = NULL
 
-	vobj <- voomWithDreamWeights(dge, form, metadata, weights=w)
+	vobj <- voomWithDreamWeights(dge2, form, metadata, weights=w, prior.count=0)
 	fit2 = dream(vobj, form, metadata)
 	fit2$targets = vobj$targets[,1:3]
 	fit2$EList = vobj
@@ -258,7 +274,8 @@ test_voomLmFit = function(){
 	fit1 = voomLmFit2( dge, dsgn, prior.weights=w  / mean(w))
 	fit1$EList$genes = NULL
 
-	vobj <- voomWithDreamWeights(dge, form, metadata, weights=w )
+	vobj <- voomWithDreamWeights(dge2, form, metadata, weights=w, prior.count=0 )
+
 	fit2 = dream(vobj, form, metadata)
 	fit2$targets = vobj$targets[,1:3]
 	fit2$EList = vobj
@@ -678,7 +695,4 @@ voomLmFit2 = function (counts, design = NULL, block = NULL, prior.weights = NULL
     }
     fit
 }
-
-
-